+++ Sources/Internal/Scene3D/Systems/EventSystem.cpp
@@ -62,13 +62,13 @@ void EventSystem::GroupNotifyAllSystems(Vector<Entity *> & entities, uint32 even
 	for(uint32 i = 0; i < size; ++i)
 	{
         SceneSystem * system = container[i];
-        uint32 requiredComponentFlags = system->GetRequiredComponents();
+        uint64 requiredComponentFlags = system->GetRequiredComponents();
 
         uint32 entityVectorSize = entities.size();
         for (uint32 k = 0; k < entityVectorSize; ++k)
         {
             Entity * entity = entities[k];
-            uint32 componentsInEntity = entity->GetAvailableComponentFlags();
+            uint64 componentsInEntity = entity->GetAvailableComponentFlags();
 
             if ((requiredComponentFlags & componentsInEntity) == requiredComponentFlags)
                 system->ImmediateEvent(entity, event);
@@ -81,11 +81,11 @@ void EventSystem::NotifyAllSystems(Entity * entity, uint32 event)
 {
 	Vector<SceneSystem*> & container = registeredSystems[event];
 	uint32 size = container.size();
-    uint32 componentsInEntity = entity->GetAvailableComponentFlags();
+    uint64 componentsInEntity = entity->GetAvailableComponentFlags();
 	for(uint32 i = 0; i < size; ++i)
 	{
         SceneSystem * system = container[i];
-        uint32 requiredComponentFlags = system->GetRequiredComponents();
+        uint64 requiredComponentFlags = system->GetRequiredComponents();
         if ((requiredComponentFlags & componentsInEntity) == requiredComponentFlags)
             system->ImmediateEvent(entity, event);
     }
